requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
ruff>=0.3.0
